+++ apps/dart_quotes/lib/data/firebase.imports.dart
@@ -1,0 +1,22 @@
+// GENERATED FILE, DO NOT MODIFY
+// Generated with jaspr_builder
+
+export '../generated/imports/_web.dart' if (dart.library.io) '../generated/imports/_stubs.dart'
+    show
+        Firebase,
+        FirebaseApp,
+        FirebaseOrStubbed,
+        FirebaseAppOrStubbed,
+        FirebaseFirestore,
+        FieldValue,
+        DocumentSnapshot,
+        FirebaseFirestoreOrStubbed,
+        FieldValueOrStubbed,
+        DocumentSnapshotOrStubbed,
+        FirebaseAuth,
+        FirebaseAuthOrStubbed,
+        DefaultFirebaseOptions,
+        DefaultFirebaseOptionsOrStubbed;
+
+export '../generated/imports/_vm.dart' if (dart.library.html) '../generated/imports/_stubs.dart'
+    show FirebaseAdminApp, Credential, FirebaseAdminAppOrStubbed, CredentialOrStubbed, Firestore, FirestoreOrStubbed;