+++ packages/jaspr/lib/server.dart
@@ -5,6 +5,7 @@ export 'package:shelf/shelf.dart' show Handler, Request, Response;
 export 'jaspr.dart' hide runApp;
 export 'src/server/adapters/element_boundary_adapter.dart';
 export 'src/server/adapters/head_scope_adapter.dart';
+export 'src/server/components/async.dart';
 export 'src/server/document.dart';
 export 'src/server/markup_render_object.dart';
 export 'src/server/run_app.dart';